requirements.txt
discord.py
spotipy
requests
yt-dlp
flask
python-dotenv
